Refs #6007. Add rmw_client.c file.
